requirements.txt
# =====================
# Core Dependencies
# =====================
python-telegram-bot==20.7
gtts==2.5.1
pydub==0.25.1
python-dotenv==1.0.0

# =====================
# Utilities
# =====================
colorama==0.4.6
requests==2.31.0
aiohttp==3.8.5
Pillow==10.0.1

# =====================
# Development / Testing (optional)
# =====================
pytest==7.4.2
black==23.9.1
flake8==6.1.0
